added search, ui update
